Created searchBar component. Navbar has a searchBar.
