+++ src/components/AppNavbar.vue
@@ -1,20 +1,13 @@
 <template>
   <va-navbar text-color="light">
     <template #center>
-      <va-navbar-item>LOGO</va-navbar-item>
       <va-navbar-item>
         <router-link :to="{ name: 'Home' }">Home</router-link>
       </va-navbar-item>
     </template>
     <template #right>
       <va-navbar-item v-if="isNotHomePage">
-        <form @submit.prevent="searchUser">
-          <va-input placeholder="username">
-            <template #appendInner>
-              <va-icon name="search" />
-            </template>
-          </va-input>
-        </form>
+        <AppSearchBar />
       </va-navbar-item>
     </template>
   </va-navbar>
@@ -23,7 +16,12 @@
 import { computed, defineComponent } from "vue";
 import { useRoute } from "vue-router";
 
+import AppSearchBar from "@/components/AppSearchBar.vue";
+
 export default defineComponent({
+  components: {
+    AppSearchBar,
+  },
   setup() {
     const route = useRoute();
     const isNotHomePage = computed(() => {
