+++ src/views/Home.vue
@@ -2,41 +2,23 @@
   <div class="layout">
     <va-card>
       <va-card-content>
-        <form @submit.prevent="searchUser"> 
-          <va-input 
-            placeholder="User Name"
-            @click-append-inner="searchUser()"
-            type="text"
-            v-model="searchUserName"
-          >
-            <template #appendInner>
-              <va-icon name="search"/>
-            </template>
-          </va-input>
-        </form>
+        <AppSearchBar />
       </va-card-content>
     </va-card>
   </div>
 </template>
 
 <script>
-import { defineComponent, ref } from "vue";
-import { useRouter } from "vue-router";
+import { defineComponent } from "vue";
+
+import AppSearchBar from "@/components/AppSearchBar.vue";
 
 export default defineComponent({
   name: 'Home',
   components: {
+    AppSearchBar
   },
   setup(){
-    const router = useRouter();
-    const searchUserName = ref("");
-    const searchUser = () => {
-      router.push('/player/' + searchUserName.value)
-    }
-    return {
-      searchUserName,
-      searchUser
-    }
   }
 });
 </script>
